Packer: Simplify jQuery UI 1.11 package
Closes gh-207
