+++ template/download/theme.html
@@ -8,9 +8,11 @@
 	</select>
 </div>
 <div class="field-pair advanced-settings">
-	<label for="theme-folder-name">Theme Folder Name:</label>
-	<input type="text" id="theme-folder-name" value="{{folderName}}" name="theme-folder-name">
-	<p id="theme-folder-name-help" class="field-help-content">This field allows you to specify a name for the theme folder in your download. This is helpful if you plan to use multiple themes on a page. It defaults to "theme".</p>
+	<div class="folder-name-area">
+		<label for="theme-folder-name">Theme Folder Name:</label>
+		<input type="text" id="theme-folder-name" value="{{folderName}}" name="theme-folder-name">
+		<p id="theme-folder-name-help" class="field-help-content">This field allows you to specify a name for the theme folder in your download. This is helpful if you plan to use multiple themes on a page. It defaults to "theme".</p>
+	</div>
 
 	<label for="scope">CSS Scope:</label>
 	<input type="text" id="scope" value="" name="scope">
